fix(ghinstall): resolve G115 gosec lint error and fix .gitignore typo
## What

Avoid the uint64-to-int cast in roundRobin.Get by keeping idx as uint64
throughout and comparing with uint64(i) in the fallback loop. Fix a
typo in .gitignore where .clause was written instead of .claude.

## Why

The gosec G115 rule flags integer overflow conversions from uint64 to
int. Since len(managers) is always small and the modulo result fits in
int, the overflow is not a real risk, but the cast is unnecessary —
using uint64 for the index directly satisfies the linter and is
semantically correct.

## Notes

- No behaviour change; slice indexing with uint64 is valid in Go.
- .gitignore typo fix is cosmetic; the pattern was never matching
  anything due to the misspelling.

+++ pkg/ghinstall/ghinstall.go
@@ -104,7 +104,7 @@ func NewRoundRobin(managers []Manager) Manager {
 }
 
 func (rr *roundRobin) Get(ctx context.Context, owner, scope, identity string) (*ghinstallation.AppsTransport, int64, error) {
-	idx := int(rr.counter.Add(1) % uint64(len(rr.managers)))
+	idx := rr.counter.Add(1) % uint64(len(rr.managers))
 
 	atr, id, err := rr.managers[idx].Get(ctx, owner, scope, identity)
 	if err == nil {
@@ -116,7 +116,7 @@ func (rr *roundRobin) Get(ctx context.Context, owner, scope, identity string) (*
 	if st, ok := status.FromError(err); ok && st.Code() == codes.NotFound {
 		clog.InfoContextf(ctx, "app not installed for %q, trying other apps", owner)
 		for i, m := range rr.managers {
-			if i == idx {
+			if uint64(i) == idx {
 				continue
 			}
 			atr, id, err = m.Get(ctx, owner, scope, identity)
